Remove casts

+++ XamlControlsGallery/ControlPages/ImplicitTransitionPage.xaml.cs
@@ -46,7 +46,7 @@ namespace AppUIBasics.ControlPages
 
             RotationRectangle.CenterPoint = new System.Numerics.Vector3((float)RotationRectangle.ActualWidth / 2, (float)RotationRectangle.ActualHeight / 2, 0f);
 
-            RotationRectangle.Rotation = (float)EnsureValueIsNumber(RotationNumberBox);
+            RotationRectangle.Rotation = EnsureValueIsNumber(RotationNumberBox);
         }
         private void ScaleButton_Click(object sender, RoutedEventArgs e)
         {
@@ -67,7 +67,7 @@ namespace AppUIBasics.ControlPages
             }
             else
             {
-                customValue = (float)EnsureValueIsNumber(ScaleNumberBox);
+                customValue = EnsureValueIsNumber(ScaleNumberBox);
             }
 
             ScaleRectangle.Scale = new Vector3(customValue);
@@ -92,7 +92,7 @@ namespace AppUIBasics.ControlPages
             }
             else
             {
-                customValue = (float)EnsureValueIsNumber(TranslationNumberBox);
+                customValue = EnsureValueIsNumber(TranslationNumberBox);
             }
 
             TranslateRectangle.Translation = new Vector3(customValue);
@@ -137,13 +137,13 @@ namespace AppUIBasics.ControlPages
             }
         }
 
-        private double EnsureValueIsNumber(NumberBox numberBox)
+        private float EnsureValueIsNumber(NumberBox numberBox)
         {
             if(double.IsNaN(numberBox.Value))
             {
                 numberBox.Value = 0;
             }
-            return numberBox.Value;
+            return (float)numberBox.Value;
         }
 
         private void ThemeButton_Click(object sender, RoutedEventArgs e)
